requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4
redis==5.2.0
celery==5.4.0
boto3==1.35.70
httpx==0.27.2
python-multipart==0.0.12
